Update 20250728 (#11)
- Updated ID data type to allow for more custom object implementations in the future
- Fixed RoadMapping bugs

+++ Assets/Scripts/ObjectScripts/Objects/ObjectData.cs
@@ -7,7 +7,7 @@ public class ObjectData : ScriptableObject
     [field: SerializeField]
     public string Name { get; private set; }
     [field: SerializeField]
-    public int ID { get; private set; }
+    public long ID { get; private set; }
     [field: SerializeField]
     public int Cost { get; private set; }
     [field: SerializeField]
